requirements.txt
altgraph==0.17.2
future==0.18.2
pefile==2022.5.30
pycryptodomex==3.14.1
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.6
pypiwin32==223
pywin32==304
pywin32-ctypes==0.2.0
